requirements.txt
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Flask==3.0.0
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
gunicorn==21.2.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
schedule==1.2.2
SQLAlchemy==2.0.42
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
